+++ auth-data/src/main/java/com/google/android/horologist/auth/data/oauth/pkce/AuthPKCEConfigRepository.kt
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.auth.data.pkce.impl.google
+package com.google.android.horologist.auth.data.oauth.pkce
 
 import com.google.android.horologist.auth.data.ExperimentalHorologistAuthDataApi
 
 @ExperimentalHorologistAuthDataApi
-public data class AuthPKCEOAuthCodeGooglePayload(
-    val code: String,
-    val redirectUrl: String
-)
+public interface AuthPKCEConfigRepository<Config> {
+
+    public suspend fun fetch(): Config
+}
