+++ auth-data/src/main/java/com/google/android/horologist/auth/data/oauth/pkce/AuthPKCETokenPayloadListener.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.auth.data.pkce
+package com.google.android.horologist.auth.data.oauth.pkce
 
 import com.google.android.horologist.auth.data.ExperimentalHorologistAuthDataApi
