+++ auth-data/src/main/java/com/google/android/horologist/auth/data/oauth/pkce/impl/google/AuthPKCETokenRepositoryGoogleImpl.kt
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.auth.data.pkce.impl.google
+package com.google.android.horologist.auth.data.oauth.pkce.impl.google
 
 import android.util.Log
 import com.google.android.horologist.auth.data.ExperimentalHorologistAuthDataApi
 import com.google.android.horologist.auth.data.oauth.common.impl.google.api.GoogleOAuthService
 import com.google.android.horologist.auth.data.oauth.common.impl.google.api.GoogleOAuthService.Companion.GRANT_TYPE_PARAM_AUTH_CODE_GRANT_VALUE
 import com.google.android.horologist.auth.data.oauth.common.impl.google.api.TokenResponse
-import com.google.android.horologist.auth.data.pkce.AuthPKCETokenRepository
-import com.google.android.horologist.auth.data.pkce.impl.AuthPKCEDefaultConfig
+import com.google.android.horologist.auth.data.oauth.pkce.AuthPKCETokenRepository
+import com.google.android.horologist.auth.data.oauth.pkce.impl.AuthPKCEDefaultConfig
 import kotlinx.coroutines.CancellationException
 
 @ExperimentalHorologistAuthDataApi
